Get Add update delete download calls are working good

+++ src/main/java/com/example/Data/dictionary/service/DataStorageService.java
@@ -68,6 +68,10 @@ public class DataStorageService {
     }
 
     public ResponseEntity<Object> deleteEntity(String dataId) {
+        boolean exits=metaDataRepository.existsByPrimaryKey(dataId);
+        if(!exits){
+            return new ResponseEntity<>("Id does not exits",HttpStatus.NOT_ACCEPTABLE);
+        }
         MetaDataModel dataEntry=metaDataRepository.findByPrimaryKey(dataId);
         metaDataRepository.delete(dataEntry);
         return new ResponseEntity<>("Deleted Successfully",HttpStatus.ACCEPTED);
